Changed employeed ID from int to long for epoch time

+++ src/model/EmployeeInterface.java
@@ -48,7 +48,7 @@ public interface EmployeeInterface {
 	/**
 	 * This is the method that will set an
 	 * employee's ID. It will take the epoch time
-	 * and store it in an integer. This value can
+	 * and store it in a long. This value can
 	 * never be changed. It's only purpose is to
 	 * serve as a key in a hashmap.
 	 */
@@ -92,6 +92,6 @@ public interface EmployeeInterface {
 	 * 
 	 * @return
 	 */
-	public int getID();
+	public long getID();
 
 }// End of the 'EmployeedInterface' class
